Adding TSDF and swap e0 and e1 in local to glo mat
The TSDF do not produce good result for now.
The fitting of cloud of points does not need to compute twice the
transform of the cloud of points. by swapping the first and the second
eigen vector in the local to global transform matrix the clouds of
points are directly correctly fitting the bounding boxes